requirements.txt
playwright==1.41.0
mcp==1.24.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pyyaml>=6.0


